Refactor StageInfo: rename BetterBurnTime to StageInfoPanel, extract BurnDurationPatch
- Rename BetterBurnTime -> StageInfoPanel (better describes what it does)
- Extract Patch_CorrectedBurnDuration into BurnDurationPatch.cs
- Update all references in Mod.cs
