+++ AdvancedFlightComputer/Features/StageInfo/StageInfoPanel.cs
@@ -23,7 +23,7 @@ namespace AdvancedFlightComputer.Features.StageInfo;
 /// Analysis results are cached and refreshed periodically (~0.5s) to avoid
 /// running StageAnalyzer every frame.
 /// </summary>
-static class BetterBurnTime
+static class StageInfoPanel
 {
     private const int UpdateIntervalFrames = 30;
 
@@ -86,7 +86,7 @@ static class BetterBurnTime
             .MakeGenericMethod(typeof(Decoupler));
 
         harmony.Patch(GameReflection.StagingWindow_DrawContent!,
-            prefix: new HarmonyMethod(typeof(BetterBurnTime), nameof(DrawContentPrefix)));
+            prefix: new HarmonyMethod(typeof(StageInfoPanel), nameof(DrawContentPrefix)));
 
         harmony.CreateClassProcessor(typeof(Patch_CorrectedBurnDuration)).Patch();
         harmony.CreateClassProcessor(typeof(StageAnalyzerDebug.Patch_AnalyzeAfterStaging)).Patch();
@@ -455,38 +455,4 @@ static class BetterBurnTime
     }
 
     #endregion
-}
-
-/// <summary>
-/// Overrides BurnDuration and IgnitionTime with multi-stage values after the
-/// background job's FlightComputer copy is applied on the main thread.
-///
-/// The game's UpdateBurnTarget (worker thread) computes a single-stage
-/// BurnDuration using only the currently active engines. This postfix
-/// replaces it with the multi-stage burn time from our StageAnalyzer,
-/// which accounts for all future stages.
-///
-/// This corrects the BURN TIME and START BURN IN gauge rollers, which
-/// read from FlightComputer.Burn on the main thread during the render pass.
-///
-/// The worker thread's ignition decision uses its own freshly-computed
-/// single-stage value (slightly off for multi-stage burns). This is
-/// acceptable because auto-staging handles burn continuation.
-/// </summary>
-[HarmonyPatch(typeof(Vehicle), "UpdateFromTaskResults")]
-static class Patch_CorrectedBurnDuration
-{
-    static void Postfix(Vehicle __instance)
-    {
-        if (__instance != Program.ControlledVehicle) return;
-
-        FlightComputer fc = __instance.FlightComputer;
-        if (fc.Burn == null) return;
-
-        float? corrected = BetterBurnTime.GetCorrectedBurnDuration();
-        if (corrected == null || corrected.Value <= 0f) return;
-
-        fc.Burn.BurnDuration = corrected.Value;
-        fc.Burn.IgnitionTime = fc.Burn.ImpulsiveInstant - 0.5 * (double)fc.Burn.BurnDuration;
-    }
 }
